Current Work In Progress
This .R file contains the most updated code that I've been working on. It transforms the x-axis to time, solves the smoothing degrees of freedom problem, and also attempts to transform the y axis to depth. The output of this script produces a recovered trace, but there are some a handful of things I need to work on / confirm before moving forward with this code. Everything is starting to come together!